requirements.txt
amqp==5.1.1
asgiref==3.3.4
async-timeout==4.0.2
autopep8==1.5.7
billiard==3.6.4.0
celery==5.2.2
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2
django-admin-footer==1.0.1
django-admin-honeypot==1.1.0
django-admin-interface==0.14.2
django-admin-logs==1.0.2
django-apptemplates==1.5
django-colorfield==0.7.2
django-defender==0.9.1
django-extensions==3.1.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-postgres-metrics==0.10.1
fitdecode==0.10.0
funcy==1.17
gpxpy==1.5.0
kombu==5.2.4
lxml==4.9.1
numpy==1.23.1
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
redis==4.3.4
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
